requirements.txt
albumentations==1.3.0
matplotlib==3.6.1
numpy==1.23.1
opencv_python_headless==4.6.0.66
pandas==1.5.0
PyYAML==6.0
scikit_learn==1.1.2
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.1
